LIBAVALON-306: Making dataprovider configurable
Created an enum class, DataProviderType, that maps the name
Fedora to the FedoraDataProvider class. This makes it so that when the
Avalon DataProvider is created later, it would just need the class
created and a keyword mapping added to the Enum class. Also updated the
method signature in app and the call to it in server.py to take in the type
configuration, and the method error will throw a Runtime exception if
the given type isn't found.

https://umd-dit.atlassian.net/browse/LIBAVALON-306