LIBAVALON-305: Adding str casting
I think this is a result of how the handles are stored differntly in the
fedora solr vs avalon solr. Fedora's handle is just a string, while in
avalon it's a list with a string inside it.

For now just casting it as a string will make sure that the app doesn't
crash when trying to list records, but I think that updating the avalon
solr to make the permalink_tesim be a string instead of a list would be
the better solution.

https://umd-dit.atlassian.net/browse/LIBAVALON-305