LIBFCREPO-1211: Fixing inclusivity with until keyword
Fixing an issue, where if the date provided to until does not have the
timestamp, the datetime created for it will default the time to 0:00:00,
which causes solr queries to not pick up dates created on that day.

Also updating the values in solr.py to correctly label them as datetimes
instead of strings.

https://umd-dit.atlassian.net/browse/LIBFCREPO-1211